// src/app/state/profile/profile.enum.ts
export enum GET_USER {
    GET_USER_PROFILE = '[GET] USER PROFILE DATA',
    GET_USER_PROFILE_SUCCESS = '[SET] USER PROFILE DATA SUCCESS',
    UPDATE_USER_PROFILE = '[UPDATE] USER PROFILE DATA',
    UPDATE_USER_PROFILE_SUCCESS = '[UPDATE] USER PROFILE DATA SUCCESS',
    GET_ERROR = '[GET] GET_ERROR',
    SET_ERROR = '[SET] GET_ERROR',
    UPDATE_USER_PASSWORD = '[SET] UPDATE_USER_PASSWORD',
    UPDATE_USER_PASSWORD_SUCCESS = '[SET] UPDATE_USER_PASSWORD_SUCCESS'
}